src/requirements.txt
ansible==6.5.0
asgiref==3.5.2
Django==4.1.3
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2-binary==2.9.4
pydantic==1.10.2
pyTelegramBotAPI==4.7.1
requests==2.28.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
telebot==0.0.4
